ENG-54868: Agent Python script

+++ web-agent/worker.py
@@ -4,19 +4,22 @@ import os
 import random
 import string
 import sys
-import requests
-import json
-import time
+import uuid
+from typing import Optional
 
+import requests
+import time
 
 api_key = None
 server_url = None
 
 letters = string.ascii_letters
 rand_string = ''.join(random.choice(letters) for _ in range(10))
-output_file = "/data/large_output_file.txt" + rand_string
+output_file_folder = '/data'
+output_file = output_file_folder + "/" + "large_output_file.txt" + rand_string
 
-max_file_size = 1024*10
+max_file_size = 1024 * 10
+
 
 def main():
     global api_key, server_url
@@ -28,8 +31,6 @@ def main():
     server_url = args.serverUrl
     api_key = args.apiKey
 
-
-
     if api_key is None:
         api_key = os.getenv("api_key")
 
@@ -39,23 +40,26 @@ def main():
         try:
             # Get the next task
             print("Requesting task...")
-            get_task_response = requests.get(f"{server_url}/api/httpTeleport/getTask", headers=headers, timeout=25)
+            get_task_response = requests.get(
+                f"{server_url}/api/httpTeleport/getTask",
+                headers=headers,
+                timeout=25)
             
             if get_task_response.status_code == 200:
-                task = get_task_response.json()['data']
+                task = get_task_response.json().get('data', None)
                 if task is None:
-                    print(f"Received empty task")
+                    print(f"Received empty task", get_task_response)
                     return
                 print(f"Received task: {task['taskId']}")
-                
+
                 # Process the task
                 result = process_task(task)
-                
+
                 # Update the task
                 update_task_response = requests.post(
                     f"{server_url}/api/httpTeleport/putResult",
-                    params={'taskId': task['taskId']},
-                    json={'result': result},
+                    headers=_get_headers(),
+                    json=task,
                     timeout=30
                 )
 
@@ -74,8 +78,10 @@ def main():
             time.sleep(5)
         finally:
             ##remove the output generated file
-            os.remove(output_file)
-            return
+            if os.path.exists(output_file):
+                os.remove(output_file)
+                return
+
 
 def _get_headers():
     headers = {
@@ -84,18 +90,18 @@ def _get_headers():
     }
     return headers
 
+
 def _validate_task(task):
     return
+
 
 def process_task(task):
     tenant = task.get('tenant')
     url = task.get('url')
     input = task.get('input')
-    taskId = task.geta('taskId')
+    taskId = task.get('taskId')
     headers = task.get('requestHeaders', {})
     method = task.get('method', 'GET').upper()
-
-
 
     print(f"Processing task {task['taskId']}: {method} {url}")
 
@@ -110,9 +116,9 @@ def process_task(task):
             if is_chunked:
                 print("Processing in chunks...")
                 # Process the response in chunks
-                for chunk in response.iter_content(chunk_size=1024*10):
+                for chunk in response.iter_content(chunk_size=1024 * 10):
                     if chunk:
-                        with open(output_file, 'a') as f:
+                        with open(output_file, 'ab') as f:
                             f.write(chunk)
 
             else:
@@ -138,7 +144,8 @@ def process_task(task):
 
     except requests.exceptions.RequestException as e:
         print(f"Error processing task {task['taskId']}: {e}")
-        return {'error': str(e)}
+        task['output'] = str(e)
+        return task
 
 
 def _update_task_with_response(task, response, s3_upload_url):
@@ -161,17 +168,28 @@ def upload_s3(preSignedUrl):
         print("Error uploading s3 signed  ")
         return False
 
-def get_s3_upload_url(taskId):
-    params = {'fileName': taskId}
+def _createFolder():
+    if not os.path.exists(output_file_folder):  # Check if the directory exists
+        try:
+            os.mkdir(output_file_folder)  # Create the directory if it doesn't exist
+        except Exception as e:
+            print(f"An error occurred while creating the folder: {e}")
+    else:
+        print(f"Directory '{output_file_folder}' already exists.")
+
+def get_s3_upload_url(taskId: str) -> Optional[str]:
+
+    params = {'fileName': taskId + str(uuid.uuid1())}
     get_s3_url = requests.get(f"{server_url}/api/httpTeleport/get-signed-url", params=params,
-                                     headers=_get_headers(), timeout=25)
+                              headers=_get_headers(), timeout=25)
 
     if get_s3_url.status_code == 200:
-        return get_s3_url.json()['data']
+        return get_s3_url.json().get('data', None)
     else:
-        raise Exception("Unable to get signedUrl " ,get_s3_url.status_code ,get_s3_url.content)
+        raise Exception("Unable to get signedUrl ", get_s3_url.status_code, get_s3_url.content)
 
 
 
 if __name__ == "__main__":
+    _createFolder()
     main()
